De-duplication around object store _construct_path.
