+++ lib/galaxy/objectstore/pithos.py
@@ -28,6 +28,7 @@ from galaxy.util import (
 from galaxy.util.path import safe_relpath
 from . import ConcreteObjectStore
 from ._util import fix_permissions
+from .caching import UsesCache
 
 NO_KAMAKI_ERROR_MESSAGE = (
     "ObjectStore configured, but no kamaki.clients dependency available."
@@ -89,7 +90,7 @@ def parse_config_xml(config_xml):
     return r
 
 
-class PithosObjectStore(ConcreteObjectStore):
+class PithosObjectStore(ConcreteObjectStore, UsesCache):
     """
     Object store that stores objects as items in a Pithos+ container.
     Cache is ignored for the time being.
@@ -152,69 +153,6 @@ class PithosObjectStore(ConcreteObjectStore):
             return
         if project and c.get("x-container-policy-project") != project:
             self.pithos.reassign_container(project)
-
-    def _construct_path(
-        self,
-        obj,
-        base_dir=None,
-        dir_only=None,
-        extra_dir=None,
-        extra_dir_at_root=False,
-        alt_name=None,
-        obj_dir=False,
-        in_cache=False,
-        **kwargs,
-    ):
-        """Construct path from object and parameters"""
-        # param extra_dir: should never be constructed from provided data but
-        # just make sure there are no shenannigans afoot
-        if extra_dir and extra_dir != os.path.normpath(extra_dir):
-            log.warning(f"extra_dir is not normalized: {extra_dir}")
-            raise ObjectInvalid("The requested object is invalid")
-        # ensure that any parent directory references in alt_name would not
-        # result in a path not contained in the directory path constructed here
-        if alt_name:
-            if not safe_relpath(alt_name):
-                log.warning(f"alt_name would locate path outside dir: {alt_name}")
-                raise ObjectInvalid("The requested object is invalid")
-            # alt_name can contain parent directory references, but S3 will not
-            # follow them, so if they are valid we normalize them out
-            alt_name = os.path.normpath(alt_name)
-        rel_path = os.path.join(*directory_hash_id(self._get_object_id(obj)))
-        if extra_dir is not None:
-            if extra_dir_at_root:
-                rel_path = os.path.join(extra_dir, rel_path)
-            else:
-                rel_path = os.path.join(rel_path, extra_dir)
-
-        # for JOB_WORK directory
-        if obj_dir:
-            rel_path = os.path.join(rel_path, str(self._get_object_id(obj)))
-        if base_dir:
-            base = self.extra_dirs.get(base_dir)
-            return os.path.join(base, rel_path)
-
-        # Pithos+ folders are marked by having trailing '/' so add it now
-        rel_path = f"{rel_path}/"
-
-        if not dir_only:
-            an = alt_name if alt_name else f"dataset_{self._get_object_id(obj)}.dat"
-            rel_path = os.path.join(rel_path, an)
-
-        if in_cache:
-            return self._get_cache_path(rel_path)
-
-        return rel_path
-
-    def _get_cache_path(self, rel_path):
-        return os.path.abspath(os.path.join(self.staging_path, rel_path))
-
-    def _in_cache(self, rel_path):
-        """Check if the given dataset is in the local cache and return True if
-        so.
-        """
-        cache_path = self._get_cache_path(rel_path)
-        return os.path.exists(cache_path)
 
     def _pull_into_cache(self, rel_path):
         # Ensure the cache directory structure exists (e.g., dataset_#_files/)
